refactor: Move `text-selection` logic to `EmojiPicker`
